Create archives module (to avoid circular imports between path and files) + remove some depr functions + fix erroneous merge
