+++ sertit/files.py
@@ -19,11 +19,7 @@ import hashlib
 import json
 import logging
 import os
-import re
 import shutil
-import tarfile
-import tempfile
-import zipfile
 from datetime import date, datetime
 from enum import Enum
 from json import JSONDecoder, JSONEncoder
@@ -32,10 +28,8 @@ from typing import Any, Union
 
 import dill
 import numpy as np
-from lxml import etree, html
-from tqdm import tqdm
 
-from sertit import AnyPath, logs, path
+from sertit import AnyPath, logs, path, s3
 from sertit.logs import SU_NAME
 from sertit.strings import DATE_FORMAT
 from sertit.types import AnyPathStrType, AnyPathType
@@ -163,454 +157,6 @@ def real_rel_path(raw_path: AnyPathStrType, start: AnyPathStrType) -> AnyPathTyp
         "This function is deprecated. Import it from 'sertit.path' instead of 'sertit.files'"
     )
     return path.real_rel_path(raw_path, start)
-
-
-def extract_file(
-    file_path: AnyPathStrType,
-    output: AnyPathStrType,
-    overwrite: bool = False,
-) -> AnyPathType:
-    """
-    Extract an archived file (zip or others). Overwrites if specified.
-    If the archive don't contain a root directory with the name of the archive without the extension, create it
-
-    Args:
-        file_path (str): Archive file path
-        output (str): Output where to put the extracted directory
-        overwrite (bool): Overwrite found extracted directory
-
-    Returns:
-        AnyPathType: Extracted directory paths
-
-    Example:
-        >>> file_path = 'D:/path/to/zip.zip'
-        >>> output = 'D:/path/to/output'
-        >>> extract_file(file_path, output, overwrite=True)
-        D:/path/to/output/zip'
-    """
-    # Convert to path
-    file_path = AnyPath(file_path)
-    output = AnyPath(output)
-
-    # In case a folder is given, returns it (this means that the file is already extracted)
-    if file_path.is_dir():
-        return file_path
-
-    # Beware with .SEN3 and .SAFE extensions
-    archive_output = output.joinpath(path.get_filename(file_path))
-
-    # In case not overwrite and the extracted directory already exists
-    if not overwrite and archive_output.exists():
-        LOGGER.debug(
-            "Already existing extracted %s. It won't be overwritten.",
-            archive_output,
-        )
-        return archive_output
-
-    def extract_sub_dir(arch, filename_list):
-        top_level_files = list({item.split("/")[0] for item in filename_list})
-
-        # When the only root directory in the archive has the right name, we don't have to create it
-        if len(top_level_files) == 1 and archive_output.name == path.get_filename(
-            top_level_files[0]
-        ):
-            arch.extractall(archive_output.parent)
-            archive_output.parent.joinpath(top_level_files[0]).rename(archive_output)
-        else:
-            arch.extractall(archive_output)
-
-    # Manage archive type
-    if file_path.suffix == ".zip":
-        with zipfile.ZipFile(file_path, "r") as zip_file:
-            extract_sub_dir(zip_file, zip_file.namelist())
-    elif file_path.suffix == ".tar" or file_path.suffixes == [".tar", ".gz"]:
-        with tarfile.open(file_path, "r") as tar_file:
-            extract_sub_dir(tar_file, tar_file.getnames())
-    elif file_path.suffix == ".7z":
-        try:
-            import py7zr
-
-            with py7zr.SevenZipFile(file_path, "r") as z7_file:
-                extract_sub_dir(z7_file, z7_file.getnames())
-        except ModuleNotFoundError as exc:
-            raise TypeError("Please install 'py7zr' to extract .7z files") from exc
-    else:
-        raise TypeError(
-            f"Only .zip, .tar, .tar.gz and .7z files can be extracted, not {file_path}"
-        )
-
-    return archive_output
-
-
-def extract_files(
-    archives: list, output: AnyPathStrType, overwrite: bool = False
-) -> list:
-    """
-    Extract all archived files. Overwrites if specified.
-
-    Example:
-        >>> file_path = ['D:/path/to/zip1.zip', 'D:/path/to/zip2.zip']
-        >>> output = 'D:/path/to/output'
-        >>> extract_files(file_path, output, overwrite=True)
-        ['D:/path/to/output.zip1', 'D:/path/to/output.zip2']
-
-    Args:
-        archives (list of str): List of archives to be extracted
-        output (str): Output folder where extracted files will be written
-        overwrite (bool): Overwrite found extracted files
-
-    Returns:
-        list: Extracted files (even pre-existing ones)
-    """
-    LOGGER.info("Extracting products in %s", output)
-    progress_bar = tqdm(archives)
-    extracts = []
-    for arch in progress_bar:
-        progress_bar.set_description(f"Extracting product {os.path.basename(arch)}")
-        extracts.append(extract_file(arch, output, overwrite))
-
-    return extracts
-
-
-def get_archived_file_list(archive_path: AnyPathStrType) -> list:
-    """
-    .. deprecated:: 1.30.0
-       Import it from :py:mod:`sertit.path` instead of :py:mod:`sertit.files`
-
-    Get the list of all the files contained in an archive.
-
-    Args:
-        archive_path (AnyPathStrType): Archive path
-
-    Returns:
-        list: All files contained in the given archive
-
-    Example:
-        >>> arch_path = 'D:/path/to/zip.zip'
-        >>> get_archived_file_list(arch_path, file_regex)
-        ['file_1.txt', 'file_2.tif', 'file_3.xml', 'file_4.geojson']
-    """
-    logs.deprecation_warning(
-        "This function is deprecated. Import it from 'sertit.path' instead of 'sertit.files'"
-    )
-    return path.get_archived_file_list(archive_path)
-
-
-def get_archived_path(
-    archive_path: AnyPathStrType, file_regex: str, as_list: bool = False
-) -> Union[list, AnyPathType]:
-    """
-    .. deprecated:: 1.30.0
-       Import it from :py:mod:`sertit.path` instead of :py:mod:`sertit.files`
-
-    Get archived file path from inside the archive.
-
-    .. WARNING::
-        If :code:`as_list` is :code:`False`, it will only return the first file matched !
-
-    You can use this `site <https://regexr.com/>`_ to build your regex.
-
-    Example:
-        >>> arch_path = 'D:/path/to/zip.zip'
-        >>> file_regex = '.*dir.*file_name'  # Use .* for any character
-        >>> path = get_archived_path(arch_path, file_regex)
-        'dir/filename.tif'
-
-    Args:
-        archive_path (AnyPathStrType): Archive path
-        file_regex (str): File regex (used by re) as it can be found in the getmembers() list
-        as_list (bool): If true, returns a list (including all found files). If false, returns only the first match
-
-    Returns:
-        Union[list, str]: Path from inside the zipfile
-    """
-    logs.deprecation_warning(
-        "This function is deprecated. Import it from 'sertit.path' instead of 'sertit.files'"
-    )
-    return path.get_archived_path(archive_path, file_regex, as_list)
-
-
-def get_archived_rio_path(
-    archive_path: AnyPathStrType, file_regex: str, as_list: bool = False
-) -> Union[list, AnyPathType]:
-    """
-    .. deprecated:: 1.30.0
-       Import it from :py:mod:`sertit.path` instead of :py:mod:`sertit.files`
-
-    Get archived file path from inside the archive, to be read with rasterio:
-
-    - :code:`zip+file://{zip_path}!{file_name}`
-    - :code:`tar+file://{tar_path}!{file_name}`
-
-
-    See `here <https://rasterio.readthedocs.io/en/latest/topics/datasets.html?highlight=zip#dataset-identifiers>`_
-    for more information.
-
-    .. WARNING::
-        It won't be readable by pandas, geopandas or xmltree !
-
-    .. WARNING::
-        If :code:`as_list` is :code:`False`, it will only return the first file matched !
-
-    You can use this `site <https://regexr.com/>`_ to build your regex.
-
-    Args:
-        archive_path (AnyPathStrType): Archive path
-        file_regex (str): File regex (used by re) as it can be found in the getmembers() list
-        as_list (bool): If true, returns a list (including all found files). If false, returns only the first match
-
-    Returns:
-        Union[list, str]: Band path that can be read by rasterio
-
-    Example:
-        >>> arch_path = 'D:/path/to/zip.zip'
-        >>> file_regex = '.*dir.*file_name'  # Use .* for any character
-        >>> path = get_archived_tif_path(arch_path, file_regex)
-        'zip+file://D:/path/to/output.zip!dir/filename.tif'
-        >>> rasterio.open(path)
-        <open DatasetReader name='zip+file://D:/path/to/output.zip!dir/filename.tif' mode='r'>
-    """
-    logs.deprecation_warning(
-        "This function is deprecated. Import it from 'sertit.path' instead of 'sertit.files'"
-    )
-    return path.get_archived_rio_path(archive_path, file_regex, as_list)
-
-
-def read_archived_file(
-    archive_path: AnyPathStrType, regex: str, file_list: list = None
-) -> bytes:
-    """
-    Read archived file (in bytes) from :code:`zip` or :code:`tar` archives.
-
-    You can use this `site <https://regexr.com/>`_ to build your regex.
-
-    Args:
-        archive_path (AnyPathStrType): Archive path
-        regex (str): Regex (used by re) as it can be found in the getmembers() list
-        file_list (list): List of files contained in the archive. Optional, if not given it will be re-computed.
-
-    Returns:
-         bytes: Archived file in bytes
-    """
-    archive_path = AnyPath(archive_path)
-
-    # Compile regex
-    regex = re.compile(regex)
-
-    # Open tar and zip XML
-    try:
-        if archive_path.suffix == ".tar":
-            with tarfile.open(archive_path) as tar_ds:
-                # file_list is not very useful for TAR files...
-                if file_list is None:
-                    tar_mb = tar_ds.getmembers()
-                    file_list = [mb.name for mb in tar_mb]
-                name = list(filter(regex.match, file_list))[0]
-                tarinfo = tar_ds.getmember(name)
-                file_str = tar_ds.extractfile(tarinfo).read()
-        elif archive_path.suffix == ".zip":
-            with zipfile.ZipFile(archive_path) as zip_ds:
-                if file_list is None:
-                    file_list = [f.filename for f in zip_ds.filelist]
-                name = list(filter(regex.match, file_list))[0]
-                file_str = zip_ds.read(name)
-
-        elif archive_path.suffix == ".tar.gz":
-            raise TypeError(
-                ".tar.gz files are too slow to read from inside the archive. Please extract them instead."
-            )
-        else:
-            raise TypeError(
-                "Only .zip and .tar files can be read from inside its archive."
-            )
-    except IndexError as exc:
-        raise FileNotFoundError(
-            f"Impossible to find file {regex} in {path.get_filename(archive_path)}"
-        ) from exc
-
-    return file_str
-
-
-def read_archived_xml(
-    archive_path: AnyPathStrType, regex: str = None, file_list: list = None, **kwargs
-) -> etree._Element:
-    """
-    Read archived XML from :code:`zip` or :code:`tar` archives.
-
-    You can use this `site <https://regexr.com/>`_ to build your regex.
-
-    Args:
-        archive_path (AnyPathStrType): Archive path
-        regex (str): XML regex (used by re) as it can be found in the getmembers() list
-        file_list (list): List of files contained in the archive. Optional, if not given it will be re-computed.
-
-    Returns:
-         etree._Element: XML file
-
-    Example:
-        >>> arch_path = 'D:/path/to/zip.zip'
-        >>> file_regex = '.*dir.*file_name'  # Use .* for any character
-        >>> read_archived_xml(arch_path, file_regex)
-        <Element LANDSAT_METADATA_FILE at 0x1c90007f8c8>
-    """
-    if regex is None:
-        logs.deprecation_warning(
-            "'xml_regex' is deprecated, please use 'regex' instead."
-        )
-        regex = kwargs.pop("xml_regex")
-
-    xml_bytes = read_archived_file(archive_path, regex=regex, file_list=file_list)
-
-    return etree.fromstring(xml_bytes)
-
-
-def read_archived_html(
-    archive_path: AnyPathStrType, regex: str, file_list: list = None
-) -> html.HtmlElement:
-    """
-    Read archived HTML from :code:`zip` or :code:`tar` archives.
-
-    You can use this `site <https://regexr.com/>`_ to build your regex.
-
-    Args:
-        archive_path (AnyPathStrType): Archive path
-        regex (str): HTML regex (used by re) as it can be found in the getmembers() list
-        file_list (list): List of files contained in the archive. Optional, if not given it will be re-computed.
-
-    Returns:
-         html._Element: HTML file
-
-    Example:
-        >>> arch_path = 'D:/path/to/zip.zip'
-        >>> file_regex = '.*dir.*file_name'  # Use .* for any character
-        >>> read_archived_html(arch_path, file_regex)
-        <Element html at 0x1c90007f8c8>
-    """
-    html_bytes = read_archived_file(archive_path, regex, file_list=file_list)
-
-    return html.fromstring(html_bytes)
-
-
-def archive(
-    folder_path: AnyPathStrType,
-    archive_path: AnyPathStrType,
-    fmt: str = "zip",
-) -> AnyPathType:
-    """
-    Archives a folder recursively.
-
-    Args:
-        folder_path (AnyPathStrType): Folder to archive
-        archive_path (AnyPathStrType): Archive path, with or without extension
-        fmt (str): Format of the archive, used by :code:`shutil.make_archive`. Choose between [zip, tar, gztar, bztar, xztar]
-
-    Returns:
-        str: Archive filename
-
-    Example:
-        >>> folder_path = 'D:/path/to/folder_to_archive'
-        >>> archive_path = 'D:/path/to/output'
-        >>> archive = archive(folder_path, archive_path, fmt="gztar")
-        'D:/path/to/output/folder_to_archive.tar.gz'
-    """
-    archive_path = AnyPath(archive_path)
-    folder_path = AnyPath(folder_path)
-
-    tmp_dir = None
-    if path.is_cloud_path(folder_path):
-        tmp_dir = tempfile.TemporaryDirectory()
-        folder_path = folder_path.download_to(tmp_dir.name)
-
-    # Shutil make_archive needs a path without extension
-    archive_base = os.path.splitext(archive_path)[0]
-
-    # Archive the folder
-    archive_fn = shutil.make_archive(
-        archive_base,
-        format=fmt,
-        root_dir=folder_path.parent,
-        base_dir=folder_path.name,
-    )
-
-    if tmp_dir is not None:
-        tmp_dir.cleanup()
-
-    return AnyPath(archive_fn)
-
-
-def add_to_zip(
-    zip_path: AnyPathStrType,
-    dirs_to_add: Union[list, AnyPathStrType],
-) -> AnyPathType:
-    """
-    Add folders to an already existing zip file (recursively).
-
-    Args:
-        zip_path (AnyPathStrType): Already existing zip file
-        dirs_to_add (Union[list, AnyPathStrType]): Directories to add
-
-    Returns:
-        AnyPathType: Updated zip_path
-
-    Example:
-        >>> zip_path = 'D:/path/to/zip.zip'
-        >>> dirs_to_add = ['D:/path/to/dir1', 'D:/path/to/dir2']
-        >>> add_to_zip(zip_path, dirs_to_add)
-        zip.zip contains 2 more folders, dir1 and dir2
-    """
-    zip_path = AnyPath(zip_path)
-
-    # If the zip is on the cloud, cache it (zipfile doesn't like cloud paths)
-    if path.is_cloud_path(zip_path):
-        zip_path = AnyPath(zip_path.fspath)
-
-    # Check if existing zipfile
-    if not zip_path.is_file():
-        raise FileNotFoundError(f"Non existing {zip_path}")
-
-    # Convert to list if needed
-    if not isinstance(dirs_to_add, list):
-        dirs_to_add = [dirs_to_add]
-
-    # Add all folders to the existing zip
-    # Forced to use ZipFile because make_archive only works with one folder and not existing zipfile
-    with zipfile.ZipFile(zip_path, "a") as zip_file:
-        progress_bar = tqdm(dirs_to_add)
-        for dir_to_add_path in progress_bar:
-            # Just to be sure, use str instead of Paths
-            if isinstance(dir_to_add_path, Path):
-                dir_to_add = str(dir_to_add_path)
-            elif path.is_cloud_path(dir_to_add_path):
-                dir_to_add = dir_to_add_path.fspath
-            else:
-                dir_to_add = dir_to_add_path
-
-            progress_bar.set_description(
-                f"Adding {os.path.basename(dir_to_add)} to {os.path.basename(zip_path)}"
-            )
-            tmp = tempfile.TemporaryDirectory()
-            if os.path.isfile(dir_to_add):
-                dir_to_add = extract_file(dir_to_add, tmp.name)
-
-            for root, _, files in os.walk(dir_to_add):
-                base_path = os.path.join(dir_to_add, "..")
-
-                # Write dir (in namelist at least)
-                zip_file.write(root, os.path.relpath(root, base_path))
-
-                # Write files
-                for file in files:
-                    zip_file.write(
-                        os.path.join(root, file),
-                        os.path.relpath(
-                            os.path.join(root, file), os.path.join(dir_to_add, "..")
-                        ),
-                    )
-
-            # Clean tmp
-            tmp.cleanup()
-
-    return zip_path
 
 
 def get_filename(file_path: AnyPathStrType, other_exts: Union[list, str] = None) -> str:
@@ -754,7 +300,7 @@ def copy(src: AnyPathStrType, dst: AnyPathStrType) -> AnyPathType:
     src = AnyPath(src)
 
     if path.is_cloud_path(src):
-        out = src.download_to(dst)
+        out = s3.download(src, dst)
     else:
         out = None
         try:
